+++ z_units/unit.py
@@ -2,9 +2,9 @@ from __future__ import annotations
 from enum import Enum
 from typing import Optional, Union, Callable, Iterable
 
-from z_unit.environment import get_env
-from z_unit.util import multi_replace
-from z_unit import constant
+from z_units.environment import get_env
+from z_units.util import multi_replace
+from z_units import constant
 
 
 class Unit: